Cloud Services QoS Data
Previous prediction methods, although quality of service advantages, but
there are some shortcomings: they predict the quality of service for a
single property, and did not consider the impact of quality of service
attributes of software and hardware resources factors and quality of
service attributes and hardware resources the correlation between
predicted and they did not consider the effect of quality of service in
the cloud computing environment. In order to more accurately predict
cloud service QoS, regardless of the underlying hardware and software
resources factors affecting the quality of service attributes in a cloud
computing environment, that is combined with the three-tier architecture
cloud computing hardware and software resources, the use of Bayesian
network model, the underlying data Cloud service QOS property resource
projections upper, in order to predict the cloud service quality of
service. Therefore, to build a good collection of data from the server
cloud platform, the underlying data infrastructure layer and the
platform layer is mainly related to the collection of hardware
resources, namely the number of processes, CPU usage, physical memory
usage, etc., at the application layer collection service response time
pelagic resources and service availability and other data. Response time
refers to the service consumer issues a message and receive a response
to the time interval between milliseconds. Availability is defined as
the service is available, is described as Ture express service is
available, is described as False indicates that the service is
unavailable.