Cloud Services QoS Data
In order to predict the QoS cloud services more accurately, Consider the
underlying software and hardware resource factors of service quality
attributes under the cloud computing environment. Combined with the
correlation between the hardware resources and quality of service
attributes of the three layer architecture, to predict the quality of
cloud services. Therefore, collect of data from the server cloud
platform, the underlying data that are related to the hardware resources
are mainly collected in the infrastructure layer and the platform layer,
such as the number of processes, CPU usage, physical memory usage, etc.
In application layer, the service response time and service availability
are collected as the upper resource data. The response time is the time
interval between the service consumers and the response time, in
milliseconds. Availability is defined as a service that can be used to
describe a service that is available for the Ture, described as a False
service that is not available.